Added ordering and switching categories

+++ web/js/taskmatrix.js
@@ -1,5 +1,15 @@
 $(document).ready(function() {
 
+    $('.task_list').sortable({ 
+        connectWith: '.task_list',
+        receive: function(event, ui) {
+            change_quad($(ui.item))
+        },
+        update: function(event, ui) {
+            update_list_order($(this).sortable('toArray').toString());
+        }
+    });
+
     //TO EDIT TASKS ALREADY ON THE MATRIX
     $('.quadrant li').each(function(item, callback) {
         register_task_item(this);
@@ -11,16 +21,16 @@ $(document).ready(function() {
         //(handle target is the surface target)
         switch($(handle.target).attr('id')) {
 
-            case 'first_quadrant':
+            case 'first_quadrant': case 'first_quadrant_list':
                 input_new_task('#first_quadrant');
             break;
-            case 'second_quadrant':
+            case 'second_quadrant': case 'second_quadrant_list':
                 input_new_task('#second_quadrant');
             break;
-            case 'third_quadrant':
+            case 'third_quadrant': case 'third_quadrant_list':
                 input_new_task('#third_quadrant');
             break;
-            case 'fourth_quadrant':
+            case 'fourth_quadrant': case 'fourth_quadrant_list':
                 input_new_task('#fourth_quadrant');
             break;
         }
@@ -32,6 +42,16 @@ $(document).ready(function() {
     }
 
 });
+
+function register_task_item(task_item) {
+
+    $(task_item).dblclick(function(handle) {
+        update_task(task_item);
+    });
+
+    //$(task_item).draggable({ connectToSortable: '.quadrant ul' });
+
+}
 
 function input_new_task(quadrant_id) {
     var list = $(quadrant_id + ' ul.task_list').first();
@@ -98,15 +118,24 @@ function save_new_task(input_box) {
     $(input_box).parent().remove();
 }
 
-function register_task_item(task_item) {
+function change_quad(task) {
+    var quad_id =  $(task).parent().parent().attr('id');
+    var item_id = $(task).attr('id');
+    var task_name = $(task).attr('task_name');
 
-    $(task_item).dblclick(function(handle) {
-        update_task(task_item);
-    });
-
+    create_new_task(task_name, quad_id, item_id, false, true);
 }
 
-function create_new_task(task_name, quadrant, task_id, skip_ajax) {
+function update_list_order(id_array) {
+
+    $.ajax({
+        type: 'POST',
+        url: '/taskmanager/ajaxsort/',
+        data: 'sortlist=' + id_array
+    });
+}
+
+function create_new_task(task_name, quadrant, task_id, skip_ajax, skip_add) {
 
     var urgent = false;
     var important = false;
@@ -146,17 +175,21 @@ function create_new_task(task_name, quadrant, task_id, skip_ajax) {
     }
 
     var new_task = $('<li></li>');
-    new_task.attr('task_name', task_name);
-    new_task.append('<strong>'+ task_name +'</strong>');
+    if(!skip_add) {
+        new_task.attr('task_name', task_name);
+        new_task.append('<strong>'+ task_name +'</strong>');
+    }
 
     if(skip_ajax) {
         var del_link = $('<a href="/taskmanager/delete?id='+task_id+'" class="delete_button">Delete</a>');
         del_link.fancybox();
 
-        new_task.attr('id', task_id);
-        new_task.append(del_link);
-        $(quad_name + ' ul.task_list').append(new_task);
-        register_task_item(new_task);
+        if(!skip_add) {
+            new_task.attr('id', task_id);
+            new_task.append(del_link);
+            $(quad_name + ' ul.task_list').append(new_task);
+            register_task_item(new_task);
+        }
     }
     else {
 
@@ -169,14 +202,16 @@ function create_new_task(task_name, quadrant, task_id, skip_ajax) {
             data: 'urgent=' + urgent + "&important=" + important + "&name=" + task_name + "&task_id=" + task_id,
             success: function(msg) {
 
-                var return_json = $.parseJSON(msg);
-                var del_link = $('<a href="/taskmanager/delete?id='+return_json.task_id+'" class="delete_button">Delete</a>');
-                del_link.fancybox();
+                if(!skip_add) {
+                    var return_json = $.parseJSON(msg);
+                    var del_link = $('<a href="/taskmanager/delete?id='+return_json.task_id+'" class="delete_button">Delete</a>');
+                    del_link.fancybox();
 
-                new_task.attr('id', return_json.task_id);
-                new_task.append(del_link);
-                $(quad_name + ' ul.task_list').append(new_task);
-                register_task_item(new_task);
+                    new_task.attr('id', return_json.task_id);
+                    new_task.append(del_link);
+                    $(quad_name + ' ul.task_list').append(new_task);
+                    register_task_item(new_task);
+                }
             }
         });
     }
